add FrontEnd content
